reuse cache
Cache is reused for same push branch. E.g. push on
- testing
- staging
- master
- 1.9

It also uses the arch in the key to make sure arm cache is not used by x64 and vice versa.
Key must use branch. This is because a cache cannot be accessed by 2 jobs at the same time